trying to add exception for fim data check on linux

+++ R/read_chkdate.R
@@ -56,18 +56,47 @@ read_chkdate <- function(urlin, xlsx) {
   if(brk == 20)
     stop("Couldn't connect to FTP site, sad face...")
 
-  # read file after success
-  html <- xml2::read_html(readChar(tmp, 1e6))
-  file.remove(tmp)
+  # fim download does not work as below on linux os
+  if(Sys.info()['sysname'] == 'Linux' & grep('floridamarine', con)){
 
-  # date of online file
-  srdate <- html %>%
-    xml2::xml_text() %>%
-    strsplit(., "[\n\r]+") %>%
-    .[[1]] %>%
-    grep(basename(urlin), ., value = TRUE) %>%
-    gsub('^(.*AM|.*PM).*$', '\\1', .) %>%
-    lubridate::mdy_hm(.)
+    srdate <- dat %>%
+      strsplit('\\n') %>%
+      .[[1]] %>%
+      grep(basename(urlin), ., value = TRUE) %>%
+      gsub(paste0(basename(urlin), '\\r'), '', .) %>%
+      gsub('^.*ftp\\s+', '', .) %>%
+      gsub('^[0-9]+', '', .)
+
+    # this tries to convert to date, year may be missing if uploaded in last six months
+    tmp <- suppressWarnings(lubridate::ymd_hm(srdate))
+
+    # if fails, assumes that year was missing because file was recent
+    if(is.na(tmp))
+      tmp <- srdate %>%
+      paste(lubridate::year(Sys.Date()), .) %>%
+      lubridate::ymd_hm(.)
+
+    srdate <- tmp
+
+  }
+
+  # windows proc
+  if(!(Sys.info()['sysname'] == 'Linux' & grep('floridamarine', con))){
+
+    # read file after success
+    html <- xml2::read_html(readChar(tmp, 1e6))
+    file.remove(tmp)
+
+    # date of online file
+    srdate <- html %>%
+      xml2::xml_text() %>%
+      strsplit(., "[\n\r]+") %>%
+      .[[1]] %>%
+      grep(basename(urlin), ., value = TRUE) %>%
+      gsub('^(.*AM|.*PM).*$', '\\1', .) %>%
+      lubridate::mdy_hm(.)
+
+  }
 
   # get date of local file
   lcdate <- file.info(xlsx)$mtime
